@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "iambic"
-version = "2.4.3"
+version = "3.0.0"
 description = "Data extraction and rendering library for Shakespearean text."
 authors = ["Sean Stewart <sean_stewart@me.com>"]
 license = "MIT"
@@ -30,26 +30,41 @@ classifiers = [
 include = ["*.md", "*.json"]
 
 [tool.poetry.dependencies]
-python = "^3.7"
-typical = "^2.1"
-ujson = "^2.0"
-markdown = "^3.2.2"
-pymdown-extensions = "^7.1"
+python = "^3.11"
+typical = "^2"
+ujson = "^2"
+markdown = "^3"
+pymdown-extensions = "^9"
 html2text = "^2020.1.16"
 tabulate = "^0.8.9"
 
-[tool.poetry.dev-dependencies]
-pytest = "^5.3"
-pytest-cov = "^2.8"
-pre-commit = "^1.20"
-flake8 = "^3.7.9"
-mkdocs-awesome-pages-plugin = "^2.2.0"
-mkdocs = "^1.1"
-mypy = "^0.761"
-mkdocs-material = "^7.0"
+[tool.poetry.group.docs.dependencies]
+mkdocs-awesome-pages-plugin = "^2"
+mkdocs = "^1"
+mkdocs-material = "^9"
 fontawesome-markdown = "^0.2.6"
-black = { version = "^20.8b1", allow-prereleases = true}
+
+[tool.poetry.group.dev.dependencies]
+pytest = "^7"
+pytest-cov = "^4"
+pre-commit = "^3"
+flake8 = "^6"
+flake8-bugbear = "^23"
+mypy = "^1"
+black = "^23"
+isort = "^5"
 
 [build-system]
 requires = ["poetry>=0.12"]
 build-backend = "poetry.masonry.api"
+
+[tool.mypy]
+mypy_path = "$MYPY_CONFIG_FILE_DIR/"
+no_namespace_packages = true
+python_version = "3.11"
+install_types = true
+non_interactive = true
+ignore_missing_imports = true
+no_strict_optional = true
+follow_imports = "silent"
+exclude = ".*tests/.*|.*docs/.*"
