@@ -1,30 +1,15 @@
-#!/usr/bin/env python
-# -*- coding: UTF-8 -*-
 from __future__ import annotations
 
 import dataclasses
-import functools
 from operator import attrgetter
-from typing import (
-    ClassVar,
-    Optional,
-    Union,
-    Tuple,
-    Mapping,
-    Type,
-    List,
-    TypeVar,
-    NewType,
-    Iterable,
-    Set,
-)
+from typing import ClassVar, Iterable, Mapping, NewType, Self, Type, TypeVar
 
 import typic
 from inflection import parameterize, titleize
 
 from iambic import roman
-from .base import NodeType
 
+from .base import NodeType
 
 __all__ = (
     "Act",
@@ -65,12 +50,11 @@ indexgetter = attrgetter("index")
 _T = TypeVar("_T")
 
 
-def sort_body(body: Iterable[_T]) -> Tuple[_T, ...]:
+def sort_body(body: Iterable[_T]) -> tuple[_T, ...]:
     return tuple(sorted(body, key=indexgetter))
 
 
-@typic.slotted
-@dataclasses.dataclass(unsafe_hash=True, order=True)
+@dataclasses.dataclass(slots=True, weakref_slot=True, unsafe_hash=True, order=True)
 class Act:
     """A representation of a single Act in a Play."""
 
@@ -78,25 +62,28 @@ class Act:
     index: int
     text: str
     num: int
-    body: "ActBodyT" = typic.field(compare=False, hash=False, default=())  # type: ignore
+    body: ActBodyT = dataclasses.field(compare=False, hash=False, default=())
+    id: NodeID = dataclasses.field(init=False)
+    col: str = dataclasses.field(init=False)
 
-    @typic.cached_property
-    def id(self) -> NodeID:
+    def __post_init__(self):
+        self.col = self._getcol()
+        self.id = self._getid()
+
+    def _getid(self) -> NodeID:
         return NodeID(parameterize(f"{self.type.lower()}-{self.col}"))
 
-    @typic.cached_property
-    def col(self):
+    def _getcol(self):
         return roman.numeral(self.num)
 
     @classmethod
-    def from_node(cls, node: "GenericNode") -> "Act":
+    def from_node(cls, node: GenericNode) -> Self:
         numeral = node.pieces[1]
         num = int(numeral) if numeral.isdigit() else roman.integer(numeral)
         return cls(index=node.index, text=node.match_text, num=num)
 
 
-@typic.slotted
-@dataclasses.dataclass(unsafe_hash=True, order=True)
+@dataclasses.dataclass(slots=True, weakref_slot=True, unsafe_hash=True, order=True)
 class Scene:
     """A representation of a single Scene in a play."""
 
@@ -104,19 +91,23 @@ class Scene:
     index: int
     text: str
     num: int
-    setting: Optional[str] = None
-    act: Optional[NodeID] = None
-    body: SceneBodyT = typic.field(compare=False, hash=False, default=())  # type: ignore
-    personae: PersonaeIDT = typic.field(compare=False, hash=False, default=())  # type: ignore
+    setting: str | None = None
+    act: NodeID | None = None
+    body: SceneBodyT = dataclasses.field(compare=False, hash=False, default=())
+    personae: PersonaeIDT = dataclasses.field(compare=False, hash=False, default=())
+    id: NodeID = dataclasses.field(init=False)
+    col: str = dataclasses.field(init=False)
 
-    @typic.cached_property
-    def id(self) -> NodeID:
+    def __post_init__(self):
+        self.col = self._getcol()
+        self.id = self._getid()
+
+    def _getid(self) -> NodeID:
         return NodeID(
             f"{self.act}-{self.type.lower()}-{roman.numeral(self.num).lower()}"
         )
 
-    @typic.cached_property
-    def col(self) -> str:
+    def _getcol(self) -> str:
         pre = self.act or ""
         if pre:
             pieces = pre.split("-")
@@ -128,7 +119,7 @@ class Scene:
         return f"{pre}{roman.numeral(self.num).lower()}"
 
     @classmethod
-    def from_node(cls, node: "GenericNode") -> "Scene":
+    def from_node(cls, node: GenericNode) -> Self:
         numeral = node.pieces[1]
         num = int(numeral) if numeral.isdigit() else roman.integer(numeral)
         setting = node.match.get("setting") or None
@@ -144,8 +135,7 @@ class Scene:
         )
 
 
-@typic.slotted
-@dataclasses.dataclass(unsafe_hash=True, order=True)
+@dataclasses.dataclass(slots=True, weakref_slot=True, unsafe_hash=True, order=True)
 class Prologue:
     """A representation of a single Prologue in a play.
 
@@ -156,11 +146,13 @@ class Prologue:
     type: ClassVar[NodeType] = NodeType.PROL
     index: int
     text: str
-    setting: Optional[str] = None
-    act: Optional[NodeID] = None
-    body: "LogueBodyT" = typic.field(compare=False, hash=False, default=())  # type: ignore
-    personae: PersonaeIDT = typic.field(compare=False, hash=False, default=())  # type: ignore
-    as_act: bool = typic.field(init=False)  # type: ignore
+    setting: str | None = None
+    act: NodeID | None = None
+    body: LogueBodyT = dataclasses.field(compare=False, hash=False, default=())
+    personae: PersonaeIDT = dataclasses.field(compare=False, hash=False, default=())
+    as_act: bool = dataclasses.field(init=False)
+    id: NodeID = dataclasses.field(init=False)
+    col: str = dataclasses.field(init=False)
 
     def __post_init__(self):
         # If the body conforms to the `ActBodyT` type, it should be treated as such.
@@ -168,14 +160,14 @@ class Prologue:
             self.body
             and isinstance(self.body[0], (Scene, Intermission, Epilogue, Prologue))
         )
+        self.col = self._getcol()
+        self.id = self._getid()
 
-    @typic.cached_property
-    def id(self) -> NodeID:
+    def _getid(self) -> NodeID:
         pre = f"{self.act}-" if self.act else ""
         return NodeID(f"{pre}{self.type.lower()}-{self.index}")
 
-    @typic.cached_property
-    def col(self):
+    def _getcol(self):
         pre = self.act or ""
         if pre:
             pieces = self.act.split("-")
@@ -186,15 +178,14 @@ class Prologue:
         return f"{pre}{self.type.title()}"
 
     @classmethod
-    def from_node(cls, node: "GenericNode") -> "Prologue":
+    def from_node(cls, node: GenericNode) -> Prologue:
         setting = node.match.get("setting") or None
         return cls(
             index=node.index, text=node.match_text, setting=setting, act=node.parent
         )
 
 
-@typic.slotted
-@dataclasses.dataclass(unsafe_hash=True, order=True)
+@dataclasses.dataclass(slots=True, weakref_slot=True, unsafe_hash=True, order=True)
 class Epilogue(Prologue):
     """A representation of a single Epilogue in a play.
 
@@ -205,8 +196,7 @@ class Epilogue(Prologue):
     type: ClassVar[NodeType] = NodeType.EPIL
 
 
-@typic.slotted
-@dataclasses.dataclass(unsafe_hash=True, order=True)
+@dataclasses.dataclass(slots=True, weakref_slot=True, unsafe_hash=True, order=True)
 class Intermission:
     """A representation of an Intermission in a play."""
 
@@ -214,24 +204,21 @@ class Intermission:
     index: int
     text: str
     act: NodeID
+    id: NodeID = dataclasses.field(init=False)
+    col: str = dataclasses.field(init=False)
 
-    @typic.cached_property
-    def id(self) -> NodeID:
-        return NodeID("intermission")
+    def __post_init__(self):
+        self.col = titleize(self.text)
+        self.id = NodeID("intermission")
 
     @classmethod
-    def from_node(cls, node: "GenericNode") -> "Intermission":
+    def from_node(cls, node: GenericNode) -> Self:
         if not node.parent:
             raise ValueError(f"Can't build {cls.__name__!r} from node: {node}")
         return cls(index=node.index, text=node.match_text, act=node.parent)
 
-    @typic.cached_property
-    def col(self):
-        return titleize(self.text)
 
-
-@typic.slotted
-@dataclasses.dataclass(unsafe_hash=True, order=True)
+@dataclasses.dataclass(slots=True, weakref_slot=True, unsafe_hash=True, order=True)
 class Persona:
     """A representation of a single character in a Play."""
 
@@ -239,25 +226,23 @@ class Persona:
     index: int
     text: str
     name: str
-    short: Optional[str] = None
+    short: str | None = None
+    id: NodeID = dataclasses.field(init=False)
+    ids: set[NodeID] = dataclasses.field(init=False)
+
+    def __post_init__(self):
+        self.id = NodeID(parameterize(self.name))
+        self.ids = {NodeID(parameterize(n.strip())) for n in self.name.split("&")}
 
     def __eq__(self, other) -> bool:
         return other.name == self.name if hasattr(other, "name") else False
-
-    @typic.cached_property
-    def id(self) -> NodeID:
-        return NodeID(parameterize(self.name))
-
-    @typic.cached_property
-    def ids(self) -> Set[NodeID]:
-        return {NodeID(parameterize(n.strip())) for n in self.name.split("&")}
 
     @property
     def is_multi(self) -> bool:
         return "&" in self.name
 
     @classmethod
-    def from_node(cls, node: "GenericNode") -> "Persona":
+    def from_node(cls, node: GenericNode) -> Self:
         return cls(
             index=node.index,
             text=node.match_text,
@@ -265,8 +250,7 @@ class Persona:
         )
 
 
-@typic.slotted
-@dataclasses.dataclass(unsafe_hash=True, order=True)
+@dataclasses.dataclass(slots=True, weakref_slot=True, unsafe_hash=True, order=True)
 class Entrance:
     """A representation of an entrance for character(s) in a Scene."""
 
@@ -275,28 +259,26 @@ class Entrance:
     text: str
     scene: NodeID
     personae: PersonaeIDT = ()
+    id: NodeID = dataclasses.field(init=False)
 
-    @typic.cached_property
-    def id(self) -> NodeID:
-        return NodeID(f"{self.scene}-{self.type.lower()}-{self.index}")
+    def __post_init__(self):
+        self.id = NodeID(f"{self.scene}-{self.type.lower()}-{self.index}")
 
     @classmethod
-    def from_node(cls, node: "GenericNode") -> "Entrance":
+    def from_node(cls, node: GenericNode) -> Self:
         if not node.parent:
             raise ValueError(f"Can't build {cls.__name__!r} from node: {node}")
         return cls(index=node.index, text=node.match_text, scene=node.parent)
 
 
-@typic.slotted
-@dataclasses.dataclass(unsafe_hash=True, order=True)
+@dataclasses.dataclass(slots=True, weakref_slot=True, unsafe_hash=True, order=True)
 class Exit(Entrance):
     """A representation of an exit for character(s) in a Scene."""
 
     type: ClassVar[NodeType] = NodeType.EXIT
 
 
-@typic.slotted
-@dataclasses.dataclass(unsafe_hash=True, order=True)
+@dataclasses.dataclass(slots=True, weakref_slot=True, unsafe_hash=True, order=True)
 class Action:
     """A representation of a stage direction related to a specific character."""
 
@@ -305,13 +287,15 @@ class Action:
     persona: NodeID
     scene: NodeID
     index: int
+    id: NodeID = dataclasses.field(init=False)
 
-    @typic.cached_property
-    def id(self) -> NodeID:
-        return NodeID(f"{self.scene}-{self.persona}-{self.type.lower()}-{self.index}")
+    def __post_init__(self):
+        self.id = NodeID(
+            f"{self.scene}-{self.persona}-{self.type.lower()}-{self.index}"
+        )
 
     @classmethod
-    def from_node(cls, node: "GenericNode") -> "Action":
+    def from_node(cls, node: GenericNode) -> Self:
         if not node.parent or not node.scene:
             raise ValueError(f"Can't build {cls.__name__!r} from node: {node}")
         return cls(
@@ -322,8 +306,7 @@ class Action:
         )
 
 
-@typic.slotted
-@dataclasses.dataclass(unsafe_hash=True, order=True)
+@dataclasses.dataclass(slots=True, weakref_slot=True, unsafe_hash=True, order=True)
 class Direction:
     """A representation of a stage direction."""
 
@@ -332,20 +315,19 @@ class Direction:
     scene: NodeID
     index: int
     stop: bool = True
+    id: NodeID = dataclasses.field(init=False)
 
-    @typic.cached_property
-    def id(self) -> NodeID:
-        return NodeID(f"{self.scene}-{self.type.lower()}-{self.index}")
+    def __post_init__(self):
+        self.id = NodeID(f"{self.scene}-{self.type.lower()}-{self.index}")
 
     @classmethod
-    def from_node(cls, node: "GenericNode") -> "Direction":
+    def from_node(cls, node: GenericNode) -> Self:
         if not node.parent:
             raise ValueError(f"Can't build {cls.__name__!r} from node: {node}")
         return cls(action=node.match_text, scene=node.parent, index=node.index)
 
 
-@typic.slotted
-@dataclasses.dataclass(unsafe_hash=True, order=True)
+@dataclasses.dataclass(slots=True, weakref_slot=True, unsafe_hash=True, order=True)
 class Dialogue:
     """A representation of a line of dialogue for a character in a scene."""
 
@@ -356,29 +338,28 @@ class Dialogue:
     index: int
     lineno: int
     linepart: int = 0
+    id: NodeID = dataclasses.field(init=False)
 
-    @typic.cached_property
-    def id(self) -> NodeID:
-        return NodeID(
+    def __post_init__(self):
+        self.id = NodeID(
             f"{self.persona}-{self.type.lower()}-{self.lineno}-{self.linepart}"
         )
 
     @classmethod
-    def from_node(cls, node: "GenericNode") -> "Dialogue":
+    def from_node(cls, node: GenericNode) -> Self:
         if None in {node.parent, node.scene, node.lineno, node.linepart}:
             raise ValueError(f"Can't build {cls.__name__!r} from node: {node}")
         return cls(
             line=node.match_text,
-            persona=node.parent,  # type: ignore
-            scene=node.scene,  # type: ignore
-            index=node.index,  # type: ignore
-            lineno=node.lineno,  # type: ignore
-            linepart=node.linepart,  # type: ignore
+            persona=node.parent,
+            scene=node.scene,
+            index=node.index,
+            lineno=node.lineno,
+            linepart=node.linepart,
         )
 
 
-@typic.slotted
-@dataclasses.dataclass(unsafe_hash=True, order=True)
+@dataclasses.dataclass(slots=True, weakref_slot=True, unsafe_hash=True, order=True)
 class Speech:
     """A representation of an unbroken piece of dialogue related to a single character."""
 
@@ -387,28 +368,30 @@ class Speech:
     scene: NodeID
     body: SpeechBodyT
     index: int
+    linerange: tuple[int, int] = dataclasses.field(init=False)
+    linepart: int = dataclasses.field(init=False)
+    num_lines: int = dataclasses.field(init=False)
+    id: NodeID = dataclasses.field(init=False)
 
     def __post_init__(self):
         self.body = sort_body(self.body)
+        self.linerange = self._getlinerange()
+        self.linepart = self._getlinepart()
+        self.num_lines = self._getnum_lines()
+        self.id = self._getid()
 
-    @typic.cached_property
-    def linerange(self) -> Tuple[int, int]:
+    def _getlinerange(self) -> tuple[int, int]:
         linenos = sorted((x.lineno for x in self.body if isinstance(x, Dialogue)))
         return linenos[0], linenos[-1]
 
-    @typic.cached_property
-    def linepart(self) -> int:
+    def _getlinepart(self) -> int:
         return next((x.linepart for x in self.body if isinstance(x, Dialogue)), 0)
 
-    @typic.cached_property
-    def num_lines(self) -> int:
+    def _getnum_lines(self) -> int:
         x, y = self.linerange
-        # line count starts at 1
-        # a range of 1 - 1 is 1 line
         return y - (x - 1)
 
-    @typic.cached_property
-    def id(self) -> NodeID:
+    def _getid(self) -> NodeID:
         uri = (
             f"{self.scene}-{self.persona}-{self.type.lower()}-"
             f"{'-'.join(map(str, self.linerange))}"
@@ -418,8 +401,7 @@ class Speech:
         return NodeID(uri)
 
 
-@typic.slotted
-@dataclasses.dataclass(unsafe_hash=True, order=True)
+@dataclasses.dataclass(slots=True, weakref_slot=True, unsafe_hash=True, order=True)
 class Metadata:
     """General information about a given play."""
 
@@ -427,14 +409,13 @@ class Metadata:
     rights: str = "Creative Commons Non-Commercial Share Alike 3.0"
     language: str = "en-GB-emodeng"
     publisher: str = "Published w/ ❤️ using iambic - https://pypi.org/project/iambic"
-    title: Optional[str] = None
-    subtitle: Optional[str] = None
+    title: str | None = None
+    subtitle: str | None = None
     edition: int = 1
     author: str = "William Shakespeare"
-    editors: Tuple[str, ...] = ()
-    tags: Tuple[str, ...] = ()
+    editors: tuple[str, ...] = ()
+    tags: tuple[str, ...] = ()
 
-    @functools.lru_cache(1)
     def asmeta(self):  # pragma: nocover
         dikt = {
             "creator": [{"type": "author", "text": self.author}],
@@ -459,25 +440,26 @@ class Metadata:
         return dikt
 
 
-@typic.slotted
-@dataclasses.dataclass(unsafe_hash=True, order=True)
+@dataclasses.dataclass(slots=True, weakref_slot=True, unsafe_hash=True, order=True)
 class Play:
     """A representation of a play in its entirety."""
 
     type: ClassVar[NodeType] = NodeType.PLAY
     body: PlayBodyT = ()
-    personae: Tuple[Persona, ...] = ()
-    meta: Metadata = typic.field(default_factory=Metadata)  # type: ignore
+    personae: tuple[Persona, ...] = ()
+    meta: Metadata = dataclasses.field(default_factory=Metadata)
+    id: NodeID = dataclasses.field(init=False)
+    linecount: int | None = dataclasses.field(init=False)
 
     def __post_init__(self):
         self.body = sort_body(self.body)
+        self.id = self._getid()
+        self.linecount = self._getlinecount()
 
-    @typic.cached_property
-    def id(self) -> NodeID:
+    def _getid(self) -> NodeID:
         return NodeID(parameterize(f"{self.meta.title}-{self.type.lower()}"))
 
-    @typic.cached_property
-    def linecount(self) -> int:
+    def _getlinecount(self) -> int:
         count = 0
         final = self.body[-1]
         scene = final.body[-1] if isinstance(final, Act) else final
@@ -487,9 +469,9 @@ class Play:
             if isinstance(entry, Speech):
                 count = entry.linerange[-1]
                 break
+
         return count
 
-    @functools.lru_cache(maxsize=1)
     def asjsonld(self):
         fn, ln = self.meta.author.split()
         data = {
@@ -530,17 +512,14 @@ class Play:
         return data
 
 
-@typic.slotted
-@dataclasses.dataclass(unsafe_hash=True, order=True)
+@dataclasses.dataclass(slots=True, weakref_slot=True, unsafe_hash=True, order=True)
 class GenericNode:
     """The root-object of a script.
 
     A script ``Node`` represents a single line of text in a script.
     """
 
-    __resolver_map__: ClassVar[
-        Mapping[NodeType, Type[ResolvedNodeT]]
-    ] = typic.FrozenDict(
+    __resolver_map__: ClassVar[Mapping[NodeType, Type[_FromNodeT]]] = typic.FrozenDict(
         {
             NodeType.ACT: Act,
             NodeType.ACTION: Action,
@@ -563,19 +542,28 @@ class GenericNode:
     text: str
     index: int
     # Additional fields which may be present
-    lineno: Optional[int] = None
-    linepart: Optional[int] = None
+    lineno: int | None = None
+    linepart: int | None = None
     # Given by text parser
     # If reading from JSON, we don't have/need this,
     # it will be provided inherently by the data-structure
     # on resolution-time.
     match: typic.FrozenDict = dataclasses.field(default_factory=typic.FrozenDict)
-    parent: Optional[NodeID] = None
-    act: Optional[NodeID] = None
-    scene: Optional[NodeID] = None
+    parent: NodeID | None = None
+    act: NodeID | None = None
+    scene: NodeID | None = None
+    pieces: list[str] = dataclasses.field(init=False)
+    match_text: str = dataclasses.field(init=False)
+    _resolved: _FromNodeT | None = dataclasses.field(
+        init=False, default=None, compare=False
+    )
 
-    @typic.cached_property
-    def resolved(self) -> "ResolvedNodeT":
+    def __post_init__(self):
+        self.match_text = self.match[self.type] if self.match else self.text
+        self.pieces = self.match_text.split()
+
+    @property
+    def resolved(self: Self) -> ResolvedNodeT:
         """Resolve a GenericNode into a typed, "resolved" Node.
 
         Raises
@@ -583,44 +571,41 @@ class GenericNode:
         TypeError
             If the NodeType provided has no resolved Node mapping.
         """
+        if self._resolved is not None:
+            return self._resolved
+
         if self.type in self.__resolver_map__:
-            model: Type[ResolvedNodeT] = self.__resolver_map__[self.type]
-            return model.from_node(self)  # type: ignore
+            model: type[_FromNodeT] = self.__resolver_map__[self.type]
+            self._resolved = model.from_node(self)
+            return self._resolved
+
         raise ValueError(
             f"Unrecognized node-type <{self.type}> for text <{self.text}>. "
             f"Valid types are: {tuple(self.__resolver_map__)}"
         )
 
-    @typic.cached_property
-    def pieces(self) -> List[str]:
-        return self.match_text.split()
 
-    @typic.cached_property
-    def match_text(self) -> str:
-        return self.match[self.type] if self.match else self.text
-
-
-ResolvedNodeT = Union[
-    Act,
-    Scene,
-    Prologue,
-    Epilogue,
-    Persona,
-    Entrance,
-    Exit,
-    Action,
-    Direction,
-    Dialogue,
-    Speech,
-    Intermission,
-]
-SpeechNodeT = Union[Dialogue, Action, Direction, Entrance, Exit]
-SpeechBodyT = Tuple[SpeechNodeT, ...]
-ActNodeT = Union[Scene, Intermission, Epilogue, Prologue]
-ActBodyT = Tuple[ActNodeT, ...]
-SceneNodeT = Union[Direction, Entrance, Exit, Speech]
-SceneBodyT = Tuple[SceneNodeT, ...]
-LogueBodyT = Union[ActBodyT, SceneBodyT]
-PlayNodeT = Union[Act, Epilogue, Prologue]
-PlayBodyT = Tuple[PlayNodeT, ...]
-PersonaeIDT = Tuple[NodeID, ...]
+_FromNodeT = (
+    Act
+    | Scene
+    | Prologue
+    | Epilogue
+    | Persona
+    | Entrance
+    | Exit
+    | Action
+    | Direction
+    | Dialogue
+    | Intermission
+)
+ResolvedNodeT = _FromNodeT | Speech
+SpeechNodeT = Dialogue | Action | Direction | Entrance | Exit
+SpeechBodyT = tuple[SpeechNodeT, ...]
+ActNodeT = Scene | Intermission | Epilogue | Prologue
+ActBodyT = tuple[ActNodeT, ...]
+SceneNodeT = Direction | Entrance | Exit | Speech
+SceneBodyT = tuple[SceneNodeT, ...]
+LogueBodyT = ActBodyT | SceneBodyT
+PlayNodeT = Act | Epilogue | Prologue
+PlayBodyT = tuple[PlayNodeT, ...]
+PersonaeIDT = tuple[NodeID, ...]
