@@ -1,5 +1,5 @@
 #!/usr/bin/env python
 # -*- coding: UTF-8 -*-
 # flake8: noqa
-from .text import parser as text
 from .data import parser as data
+from .text import parser as text
