@@ -5,8 +5,8 @@ import functools
 from markdown import Markdown
 
 from iambic import ast
-from .markdown import render_markdown
 
+from .markdown import render_markdown
 
 EXTENSIONS = ("meta", "attr_list", "pymdownx.details", "tables", "pymdownx.tabbed")
 
